requirements.txt
streamlit
plotly
sentence-transformers
faiss-cpu
torch
transformers
pandas
gradio>=4.0.0
python-dotenv
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.0.0